// src/types/supabase.ts
export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      gigs: {
        Row: {
          budget: number | null
          created_at: string
          deadline: string | null
          description: string | null
          id: number
          location: string | null
          skills_required: string[] | null
          title: string
          type: Database["public"]["Enums"]["gig_type"] | null
          user_id: string
        }
        Insert: {
          budget?: number | null
          created_at?: string
          deadline?: string | null
          description?: string | null
          id?: never
          location?: string | null
          skills_required?: string[] | null
          title: string
          type?: Database["public"]["Enums"]["gig_type"] | null
          user_id: string
        }
        Update: {
          budget?: number | null
          created_at?: string
          deadline?: string | null
          description?: string | null
          id?: never
          location?: string | null
          skills_required?: string[] | null
          title?: string
          type?: Database["public"]["Enums"]["gig_type"] | null
          user_id?: string
        }
        Relationships: []
      }
// ...
      proposals: {
        Row: {
          created_at: string
          gig_id: number
          id: number
          message: string | null
          rate: number | null
          status: Database["public"]["Enums"]["proposal_status"]
          timeline: string | null
          user_id: string
        }
        Insert: {
          created_at?: string
          gig_id: number
          id?: never
          message?: string | null
          rate?: number | null
          status?: Database["public"]["Enums"]["proposal_status"]
          timeline?: string | null
          user_id: string
        }
        Update: {
          created_at?: string
          gig_id?: number
          id?: never
          message?: string | null
          rate?: number | null
          status?: Database["public"]["Enums"]["proposal_status"]
          timeline?: string | null
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "proposals_gig_id_fkey"
            columns: ["gig_id"]
            isOneToOne: false
            referencedRelation: "gigs"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never